+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Data
 @Builder
 public class ItemDto {
-    private Long id;
+    private Integer id;
     @NotBlank(message = "Имя не должно быть пустым")
     private String name;
     @NotBlank(message = "Описание не должно быть пустым")
@@ -26,7 +26,7 @@ public class ItemDto {
     @Data
     @Builder
     public static class ItemBooking {
-        private Long id;
-        private Long bookerId;
+        private Integer id;
+        private Integer bookerId;
     }
 }
